+++ app/ui/view/molecule/drawer/drawer.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 import Drawer from './drawer';
-import { DIALOG_KEY, DialogKey } from '@/app/utils/key/dialog.key';
+import { DIALOG_KEY, DialogKey } from '@/app/utils/key/dialog-key.util';
 
 const meta = {
   title: 'ui/view/molecule/Drawer',
